refactor: shift the language selector into its own component inside the Navigation.tsx
